# educa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Educa LRS REST API Backend",
        "description": "Educa LRS API Backend",
        "contact": {
            "email": "support@digitallearning.gmbh"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/verbs": {
            "get": {
                "tags": [
                    "Verbs"
                ],
                "summary": "List all verbs",
                "operationId": "614da538305061d76a85db8434cf25f6",
                "responses": {
                    "200": {
                        "description": "List of verbs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "completed"
                                            },
                                            "iri": {
                                                "type": "string",
                                                "example": "http://adlnet.gov/expapi/verbs/completed"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Verbs"
                ],
                "summary": "Create a new verb",
                "operationId": "389d491f88bb36478be597d9a5c66e73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "iri"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "completed"
                                    },
                                    "iri": {
                                        "type": "string",
                                        "example": "http://adlnet.gov/expapi/verbs/completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Verb created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "iri": {
                                            "type": "string",
                                            "example": "http://adlnet.gov/expapi/verbs/completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/verbs/{id}": {
            "get": {
                "tags": [
                    "Verbs"
                ],
                "summary": "Retrieve a verb by ID",
                "operationId": "47d4088069b5def5b74d1d0ebdb95f43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the verb to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verb retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "completed"
                                        },
                                        "iri": {
                                            "type": "string",
                                            "example": "http://adlnet.gov/expapi/verbs/completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Verbs"
                ],
                "summary": "Update a verb",
                "operationId": "cb1d3154db673f6b41e67a25e0d8d1b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the verb to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "updated-completed"
                                    },
                                    "iri": {
                                        "type": "string",
                                        "example": "http://adlnet.gov/expapi/verbs/updated-completed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Verb updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "updated-completed"
                                        },
                                        "iri": {
                                            "type": "string",
                                            "example": "http://adlnet.gov/expapi/verbs/updated-completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Verbs"
                ],
                "summary": "Delete a verb",
                "operationId": "47b91c8fed88df533432a82d4468d90a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the verb to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verb deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Verb deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}